% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_my_stan_model2}
\alias{fit_my_stan_model2}
\title{Fit My Stan Model2}
\usage{
fit_my_stan_model2(data, ...)
}
\arguments{
\item{data}{A named list of data as expected by my_stan_model.stan}

\item{...}{Other arguments passed to rstan::sampling (e.g. iter, chains)}
}
\value{
A stanfit object
}
\description{
This function runs MCMC via rstan::sampling on the precompiled Stan model.
}
