% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyStanPkg1-package.R
\docType{package}
\name{MyStanPkg1-package}
\alias{MyStanPkg1-package}
\alias{MyStanPkg1}
\title{The 'MyStanPkg1' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
